remove User model from controller
